Add more metrics
